+++ opentelemetry-sdk/tests/logs/test_multi_log_processor.py
@@ -22,7 +22,7 @@ from abc import ABC, abstractmethod
 from unittest.mock import Mock
 
 from opentelemetry._logs import SeverityNumber
-from opentelemetry.sdk._logs import (
+from opentelemetry.sdk._logs._internal import (
     ConcurrentMultiLogRecordProcessor,
     LoggerProvider,
     LoggingHandler,
